CollectionAccessService returns boolean
- We're intending to make the CollectionAccessService "dumb" by not
relying on the CollectionRepository

+++ video-service/src/main/kotlin/com/boclips/videos/service/application/collection/DeleteCollection.kt
@@ -1,10 +1,11 @@
 package com.boclips.videos.service.application.collection
 
+import com.boclips.videos.service.application.collection.exceptions.CollectionAccessNotAuthorizedException
+import com.boclips.videos.service.application.getCurrentUserId
 import com.boclips.videos.service.domain.model.collection.CollectionId
 import com.boclips.videos.service.domain.model.collection.CollectionRepository
 import com.boclips.videos.service.domain.service.collection.CollectionAccessService
 import com.boclips.videos.service.domain.service.collection.CollectionSearchService
-import com.boclips.videos.service.domain.service.collection.CollectionService
 
 class DeleteCollection(
     private val collectionRepository: CollectionRepository,
@@ -12,7 +13,9 @@ class DeleteCollection(
     private val collectionAccessService: CollectionAccessService
 ) {
     operator fun invoke(collectionId: String) {
-        collectionAccessService.getOwnedCollectionOrThrow(collectionId)
+        if (!collectionAccessService.hasWriteAccess(collectionId)) {
+            throw CollectionAccessNotAuthorizedException(getCurrentUserId(), collectionId)
+        }
 
         collectionRepository.delete(CollectionId(collectionId))
